Fix bug where record x; stop would break the poker table
record x; stop resets all entities with models on the client and is used
extensively by Outfitter/Wardrobe to refresh client models.
This causes the OnRemove method to be fired, which we don't want.
Handles the special case where OnRemove is called via fullUpdate,
preventing the HudDraw hook from being removed and resetting the
entities on the table to their original positions.